New files and small fixes
Added a new Y-Style vertical carriage design, in two parts. The main
carriage and the bracket for the effector arms separate to allow the
parts to be printed. These make use of the new 'w' universal roller.

Added a dimensioned drawing in pdf format for making the outer braces
from 1" x .125" aluminum bar. Available at most hardware stores in the
USA.

Added some material to the Outer Brace Bottom Bracket to improve the
connection to the above mentioned aluminum bars.